Create Queue_Operations.java
A queue is an object that represents a data structure designed to have the element inserted at the end of the queue, and the element removed from the beginning of the queue.

Java.Util.Queue contains multiple elements before the process. The order of elements of the queue in Java is FIFO (first-in-first-out). It provides additional operations such as insertion, inspection, and deletion.